add table body attributes to the table

add table body attributes to the table

rtf_body(
  tbl,
  colheader = TRUE,
  page_width = 8.5,
  page_height = 11,
  orientation = "portrait",
  doctype = "wma",
  border_left = "single",
  border_right = "single",
  border_top = NULL,
  border_bottom = "double",
  border_color_left = NULL,
  border_color_right = NULL,
  border_color_top = NULL,
  border_color_bottom = NULL,
  border_width = 15,
  col_rel_width = NULL,
  col_total_width = page_width/1.4,
  cell_height = 0.15,
  cell_justification = "c",
  text_font = 1,
  text_format = NULL,
  text_color = NULL,
  text_background_color = NULL,
  text_justification = "c",
  text_font_size = 9,
  text_space_before = 15,
  text_space_after = 15,
  page_num = NULL,
  page_by = NULL,
  new_page = FALSE,
  last_row = TRUE
)

Arguments

tbl
A data frame

colheader
A boolean value to indicate whether to add default column header to the table

page_width
page width in inches

page_height
page height in inches

orientation
Orientation in 'portrait' or 'landscape'

doctype
doctype in 'csr', 'wma', or 'wmm'

border_left
left border type

border_right
right border type

border_top
top border type

border_bottom
bottom border type

border_color_left
left border color

border_color_right
right border color

border_color_top
top border color

border_color_bottom
bottom border color

border_width
border width in twips

col_rel_width
column relative width in a vector eg. c(2,1,1) refers to 2:1;1

col_total_width
column total width for the table

cell_height
height for cell in twips

cell_justification
justification for cell

text_font
text font type

text_format
text format

text_color
text color

text_background_color
text background color

text_justification
justification for text

text_font_size
text font size

text_space_before
line space before text

text_space_after
line space after text

page_num
number of rows in each page

page_by
column names to group by table in sections

new_page
a boolean value to indicate whether to separate grouped table into pages by sections

last_row
a boolean value to indicate whether the table contains the last row of the final table

Value

the same data frame tbl with additional attributes for table body

Examples

library(dplyr) # required to run examples

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

data(tbl_1)
tbl_1 %>%
  rtf_body(col_rel_width = c(3,1,3,1,3,1,3,5),
           text_justification = c("l",rep("c",7)),
           last_row = FALSE) %>%
  attributes()

#> $class
#> [1] "tbl"        "data.frame"
#> 
#> $row.names
#> [1] 1 2
#> 
#> $names
#> [1] "Trt"   "N1"    "Mean1" "N2"    "Mean2" "N3"    "Mean3" "CI"   
#> 
#> $boxh_df
#>               Trt    N1 Mean1    N2 Mean2    N3 Mean3   CI
#> group_label  <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA> <NA>
#> column_label <NA>  <NA>  <NA>  <NA>  <NA>  <NA>  <NA> <NA>
#> column_align left right  left right  left right  left left
#> 
#> $stub_df
#>   groupname rowname
#> 1      <NA>    <NA>
#> 2      <NA>    <NA>
#> 
#> $footnotes_df
#> # A tibble: 1 x 6
#>   locname     locnum grpname colname rownum text                                
#>   <chr>        <dbl> <chr>   <chr>    <int> <chr>                               
#> 1 columns_co…      4 NA      CI          NA "$^\\dagger$Based on an ANCOVA mode…
#> 
#> $styles_df
#> # A tibble: 0 x 6
#> # … with 6 variables: locname <chr>, locnum <int>, grpname <chr>,
#> #   colname <chr>, rownum <int>, text <chr>
#> 
#> $rows_df
#> # A tibble: 2 x 1
#>   rownums_start
#>           <int>
#> 1             1
#> 2             2
#> 
#> $cols_df
#> # A tibble: 8 x 1
#>   colnames_start
#>   <chr>         
#> 1 Trt           
#> 2 N1            
#> 3 Mean1         
#> 4 N2            
#> 5 Mean2         
#> 6 N3            
#> 7 Mean3         
#> 8 CI            
#> 
#> $col_labels
#> $col_labels$Trt
#> [1] "Treatment"
#> 
#> $col_labels$N1
#> [1] "N"
#> 
#> $col_labels$Mean1
#> [1] "Mean (SD)"
#> 
#> $col_labels$N2
#> [1] "N"
#> 
#> $col_labels$Mean2
#> [1] "Mean (SD)"
#> 
#> $col_labels$N3
#> [1] "N"
#> 
#> $col_labels$Mean3
#> [1] "Mean (SD)"
#> 
#> $col_labels$CI
#> [1] "$LS Mean (95% CI)\\dagger$"
#> 
#> 
#> $grp_labels
#> $grp_labels$Trt
#> [1] NA
#> 
#> $grp_labels$N1
#> [1] NA
#> 
#> $grp_labels$Mean1
#> [1] NA
#> 
#> $grp_labels$N2
#> [1] NA
#> 
#> $grp_labels$Mean2
#> [1] NA
#> 
#> $grp_labels$N3
#> [1] NA
#> 
#> $grp_labels$Mean3
#> [1] NA
#> 
#> $grp_labels$CI
#> [1] NA
#> 
#> 
#> $arrange_groups
#> $arrange_groups$groups
#> character(0)
#> 
#> 
#> $data_df
#>          Trt N1       Mean1 N2       Mean2 N3       Mean3                   CI
#> 1 Study Drug 61 16.6 (4.41) 61 -6.6 (5.95) 61 -7.0 (9.16)  -7.0 (-8.58, -5.38)
#> 2    Placebo 70 18.4 (6.34) 70 -9.0 (7.04) 70 -8.7 (8.54) -8.7 (-10.17, -7.18)
#> 
#> $opts_df
#>                             parameter  scss          category      value
#> 1                     container_width FALSE         container       auto
#> 2                    container_height FALSE         container       auto
#> 3                container_overflow_x FALSE         container       auto
#> 4                container_overflow_y FALSE         container       auto
#> 5                            table_id FALSE             table brsewjfdtk
#> 6                         table_width  TRUE             table       auto
#> 7                     table_font_size  TRUE             table       16px
#> 8              table_background_color  TRUE             table    #FFFFFF
#> 9                         margin_left  TRUE             table       auto
#> 10                       margin_right  TRUE             table       auto
#> 11             table_border_top_style  TRUE             table      solid
#> 12             table_border_top_width  TRUE             table        2px
#> 13             table_border_top_color  TRUE             table    #A8A8A8
#> 14           heading_background_color  TRUE           heading       <NA>
#> 15            heading_title_font_size  TRUE           heading       125%
#> 16         heading_subtitle_font_size  TRUE           heading        85%
#> 17        heading_border_bottom_style  TRUE           heading      solid
#> 18        heading_border_bottom_width  TRUE           heading        2px
#> 19        heading_border_bottom_color  TRUE           heading    #A8A8A8
#> 20     column_labels_background_color  TRUE           columns       <NA>
#> 21            column_labels_font_size  TRUE           columns       16px
#> 22          column_labels_font_weight  TRUE           columns    initial
#> 23               column_labels_hidden FALSE           columns      FALSE
#> 24         row_group_background_color  TRUE         row_group       <NA>
#> 25                row_group_font_size  TRUE         row_group       16px
#> 26              row_group_font_weight  TRUE         row_group    initial
#> 27         row_group_border_top_style  TRUE         row_group      solid
#> 28         row_group_border_top_width  TRUE         row_group        2px
#> 29         row_group_border_top_color  TRUE         row_group    #A8A8A8
#> 30      row_group_border_bottom_style  TRUE         row_group      solid
#> 31      row_group_border_bottom_width  TRUE         row_group        2px
#> 32      row_group_border_bottom_color  TRUE         row_group    #A8A8A8
#> 33        table_body_border_top_style  TRUE        table_body      solid
#> 34        table_body_border_top_width  TRUE        table_body        2px
#> 35        table_body_border_top_color  TRUE        table_body    #A8A8A8
#> 36     table_body_border_bottom_style  TRUE        table_body      solid
#> 37     table_body_border_bottom_width  TRUE        table_body        2px
#> 38     table_body_border_bottom_color  TRUE        table_body    #A8A8A8
#> 39                        row_padding  TRUE               row        8px
#> 40          row_striping_include_stub FALSE               row      FALSE
#> 41    row_striping_include_table_body FALSE               row       TRUE
#> 42       summary_row_background_color  TRUE       summary_row       <NA>
#> 43                summary_row_padding  TRUE       summary_row        8px
#> 44         summary_row_text_transform  TRUE       summary_row    inherit
#> 45 grand_summary_row_background_color  TRUE grand_summary_row       <NA>
#> 46          grand_summary_row_padding  TRUE grand_summary_row        8px
#> 47   grand_summary_row_text_transform  TRUE grand_summary_row    inherit
#> 48                       footnote_sep FALSE          footnote     <br />
#> 49                     footnote_glyph FALSE          footnote    numbers
#> 50                 footnote_font_size  TRUE          footnote        90%
#> 51                   footnote_padding  TRUE          footnote        4px
#> 52               sourcenote_font_size  TRUE        sourcenote        90%
#> 53                 sourcenote_padding  TRUE        sourcenote        4px
#> 
#> $formats
#> list()
#> 
#> $transforms
#> list()
#> 
#> $heading
#> $heading$title
#> [1] "ANCOVA of Change from Baseline at Week 8"
#> 
#> $heading$subtitle
#> [1] "Missing Data Approach \n Analysis Population"
#> 
#> 
#> $page_width
#> [1] 8.5
#> 
#> $page_height
#> [1] 11
#> 
#> $orientation
#> [1] "portrait"
#> 
#> $doctype
#> [1] "wma"
#> 
#> $rtf_colheader
#> $rtf_colheader$col_rel_width
#> $rtf_colheader$col_rel_width[[1]]
#> [1] NA
#> 
#> 
#> $rtf_colheader$colheader
#> $rtf_colheader$colheader[[1]]
#>    X1 X2    X3 X4    X5 X6    X7 X8
#> 1 Trt N1 Mean1 N2 Mean2 N3 Mean3 CI
#> 
#> 
#> $rtf_colheader$border_left
#> $rtf_colheader$border_left[[1]]
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]    
#> [1,] "single" "single" "single" "single" "single" "single" "single" "single"
#> 
#> 
#> $rtf_colheader$border_right
#> $rtf_colheader$border_right[[1]]
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]    
#> [1,] "single" "single" "single" "single" "single" "single" "single" "single"
#> 
#> 
#> $rtf_colheader$border_top
#> $rtf_colheader$border_top[[1]]
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]    
#> [1,] "single" "single" "single" "single" "single" "single" "single" "single"
#> 
#> 
#> $rtf_colheader$border_bottom
#> $rtf_colheader$border_bottom[[1]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] ""   ""   ""   ""   ""   ""   ""   ""  
#> 
#> 
#> $rtf_colheader$border_color_left
#> $rtf_colheader$border_color_left[[1]]
#> NULL
#> 
#> 
#> $rtf_colheader$border_color_right
#> $rtf_colheader$border_color_right[[1]]
#> NULL
#> 
#> 
#> $rtf_colheader$border_color_bottom
#> $rtf_colheader$border_color_bottom[[1]]
#> NULL
#> 
#> 
#> $rtf_colheader$border_width
#> $rtf_colheader$border_width[[1]]
#> [1] 15
#> 
#> 
#> $rtf_colheader$cell_justification
#> $rtf_colheader$cell_justification[[1]]
#> [1] "c"
#> 
#> 
#> $rtf_colheader$page_width
#> $rtf_colheader$page_width[[1]]
#> [1] 8.5
#> 
#> 
#> $rtf_colheader$col_total_width
#> $rtf_colheader$col_total_width[[1]]
#> [1] 6.071429
#> 
#> 
#> $rtf_colheader$cell_height
#> $rtf_colheader$cell_height[[1]]
#> [1] 0.15
#> 
#> 
#> $rtf_colheader$text_justification
#> $rtf_colheader$text_justification[[1]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] "c"  "c"  "c"  "c"  "c"  "c"  "c"  "c" 
#> 
#> 
#> $rtf_colheader$text_font
#> $rtf_colheader$text_font[[1]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]    1    1    1    1    1    1    1    1
#> 
#> 
#> $rtf_colheader$text_format
#> $rtf_colheader$text_format[[1]]
#> NULL
#> 
#> 
#> $rtf_colheader$text_color
#> $rtf_colheader$text_color[[1]]
#> NULL
#> 
#> 
#> $rtf_colheader$text_background_color
#> $rtf_colheader$text_background_color[[1]]
#> NULL
#> 
#> 
#> $rtf_colheader$text_font_size
#> $rtf_colheader$text_font_size[[1]]
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]    9    9    9    9    9    9    9    9
#> 
#> 
#> $rtf_colheader$text_space_before
#> $rtf_colheader$text_space_before[[1]]
#> [1] 15
#> 
#> 
#> $rtf_colheader$text_space_after
#> $rtf_colheader$text_space_after[[1]]
#> [1] 15
#> 
#> 
#> $rtf_colheader$first_row
#> $rtf_colheader$first_row[[1]]
#> [1] FALSE
#> 
#> 
#> 
#> $border_top
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]    
#> [1,] "single" "single" "single" "single" "single" "single" "single" "single"
#> [2,] ""       ""       ""       ""       ""       ""       ""       ""      
#> 
#> $border_left
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]    
#> [1,] "single" "single" "single" "single" "single" "single" "single" "single"
#> [2,] "single" "single" "single" "single" "single" "single" "single" "single"
#> 
#> $border_right
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]    
#> [1,] "single" "single" "single" "single" "single" "single" "single" "single"
#> [2,] "single" "single" "single" "single" "single" "single" "single" "single"
#> 
#> $border_bottom
#>      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]    
#> [1,] "double" "double" "double" "double" "double" "double" "double" "double"
#> [2,] "double" "double" "double" "double" "double" "double" "double" "double"
#> 
#> $border_width
#> [1] 15
#> 
#> $col_total_width
#> [1] 6.071429
#> 
#> $cell_height
#> [1] 0.15
#> 
#> $cell_justification
#> [1] "c"
#> 
#> $col_rel_width
#> [1] 3 1 3 1 3 1 3 5
#> 
#> $text_font
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]    1    1    1    1    1    1    1    1
#> [2,]    1    1    1    1    1    1    1    1
#> 
#> $text_font_size
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,]    9    9    9    9    9    9    9    9
#> [2,]    9    9    9    9    9    9    9    9
#> 
#> $text_justification
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] "l"  "c"  "c"  "c"  "c"  "c"  "c"  "c" 
#> [2,] "l"  "c"  "c"  "c"  "c"  "c"  "c"  "c" 
#> 
#> $text_space_before
#> [1] 15
#> 
#> $text_space_after
#> [1] 15
#> 
#> $page_num
#> [1] 42
#> 
#> $new_page
#> [1] FALSE
#> 
#> $pageby_db
#>          Trt N1       Mean1 N2       Mean2 N3       Mean3                   CI
#> 1 Study Drug 61 16.6 (4.41) 61 -6.6 (5.95) 61 -7.0 (9.16)  -7.0 (-8.58, -5.38)
#> 2    Placebo 70 18.4 (6.34) 70 -9.0 (7.04) 70 -8.7 (8.54) -8.7 (-10.17, -7.18)
#> 
#> $pageby_newpage_index
#> [1] 1
#> 
#> $last_row
#> [1] FALSE
#>