Add Column Header Attributes to Table

Add Column Header Attributes to Table

rtf_colheader(
  tbl,
  colheader = NULL,
  col_rel_width = NULL,
  border_left = "single",
  border_right = "single",
  border_top = "single",
  border_bottom = "",
  border_color_left = NULL,
  border_color_right = NULL,
  border_color_top = NULL,
  border_color_bottom = NULL,
  border_width = 15,
  cell_height = 0.15,
  cell_justification = "c",
  text_font = 1,
  text_format = NULL,
  text_font_size = 9,
  text_color = NULL,
  text_background_color = NULL,
  text_justification = "c",
  text_indent_first = 0,
  text_indent_left = 0,
  text_indent_right = 0,
  text_space = 1,
  text_space_before = 15,
  text_space_after = 15,
  text_convert = TRUE
)

Arguments

tbl
A data frame. Required argument.

colheader
A character string that uses " | " to separate column names. Default is NULL for a blank column header.

col_rel_width
A numeric vector for column relative width, e.g. c(2,1,1) refers to 2:1:1. Default is NULL for equal column width.

border_left
A character vector for left border type. Default is "single". All possible input can be found in r2rtf:::border_type().

border_right
A character vector for right border type. Default is "single". All possible input can be found in r2rtf:::border_type().

border_top
A character vector for top border type. Default is NULL. If it is the first row in a table for this page, the border is set to "double" otherwise the border is set to "single".

border_bottom
A character vector for bottom border type. Default is NULL for no border. All possible input can be found in r2rtf:::border_type().

border_color_left
A character vector for left border color. Default is NULL for black. All possible input can be found in grDevices::colors().

border_color_right
A character vector for right border color. Default is NULL for black. All possible input can be found in grDevices::colors().

border_color_top
A character vector for top border color. Default is NULL for black. All possible input can be found in grDevices::colors().

border_color_bottom
A character vector for bottom border color. Default is NULL for black. All possible input can be found in grDevices::colors().

border_width
A numeric value for border width in twips. Default is 15.

cell_height
A numeric value for height of cell in inch. Default is 0.15 inch.

cell_justification
A character vector for justification of cell. Default is "c" for center justification. All possible input can be found in r2rtf:::justification().

text_font
A numeric value for text font type. Default is 1 for Times New Roman. All possible input can be found in r2rtf:::font_type().

text_format
A character vector for text format. Default is NULL for normal format. Multiple format type can be assigned, e.g. "bi" will give the text both bold and italics format. All possible input can be found in r2rtf:::font_format().

text_font_size
A numeric value for text font size. Default is 9.

text_color
A character vector for text color. Default is NULL for black. All possible input can be found in grDevices::colors().

text_background_color
A character vector for text background color. Default is NULL for black. All possible input can be found in grDevices::colors().

text_justification
A character vector for text justification. Default is "c" for center justification. All possible input can be found in r2rtf:::justification().

text_indent_first
a value of first indent

text_indent_left
a value of left indent

text_indent_right
a value of right indent

text_space
a value of text space

text_space_before
Line space before text in twips. Default is 15 for 0.0104 inch.

text_space_after
Line space after text in twips. Default is 15 for 0.0104 inch.

text_convert
a logical value to convert special characters. Default is TRUE.

Value

The same data frame tbl with additional attributes for table column header.

Specification

The contents of this section are shown in PDF user manual only.

Examples

library(dplyr) # required to run examples
data(tbl_1)
tbl_1 %>%
  rtf_colheader(colheader = "Treatment | N | Mean (SD) | N | Mean (SD) | N |
                  Mean (SD) | LS Mean (95% CI)\\dagger",
                text_format = c("b", "", "u", "", "u", "","u", "i") ) %>%
  attr("rtf_colheader")

#> [[1]]
#>          X1 X2        X3 X4        X5 X6        X7                       X8
#> 1 Treatment  N Mean (SD)  N Mean (SD)  N Mean (SD) LS Mean (95% CI)\\dagger
#>