Convert RTF file to Other Format

This is an experimental function.

Usage

rtf_convert_format(
  input,
  output_file = NULL,
  output_dir = ".",
  format = "pdf",
  overwrite = FALSE
)

Arguments

input

A vector of file paths for the input file to be converted.

output_file

A vector of filename for the output file. Default is the same filename for input.

output_dir

The output directory for the converted output_dir.

format

Converted file format extension. Currently support "pdf" or "docx"

overwrite

logical; should existing destination files be overwritten?

Value

A vector of file paths for the converted files.

Details

Convert RTF files to PDF or DOCX files. Require libreoffice.