Paragraph to RTF Encode

Paragraph to RTF Encode

rtf_paragraph(
  text,
  justification = "c",
  indent_first = 0,
  indent_left = 0,
  indent_right = 0,
  space = 1,
  space_before = 180,
  space_after = 180,
  new_page = FALSE,
  hyphenation = TRUE,
  cell = FALSE
)

Arguments

text
rtf text obtained using rtf_text() function.

justification
Justification for text.

indent_first
First indent.

indent_left
Left indent.

indent_right
Right indent.

space
Paragraph space.

space_before
Line space before text.

space_after
Line space after text.

new_page
A boolean value to indicate whether to start a new page.

hyphenation
A boolean value to indicate whether to use hyphenation.

cell
A boolean value to indicate if paragraph is in table cell.

Specification

The contents of this section are shown in PDF user manual only.