Convert a UTF-8 Encoded Character String to a RTF Encoded String

Convert a UTF-8 Encoded Character String to a RTF Encoded String

Usage

utf8Tortf(text)

Arguments

text

A string to be converted.

If the unicode of a character is less than 128 (including all character on a keyboard), the character is as is. If the unicode of a character is larger or equal to 128, the character will be encoded.

Specification

The contents of this section are shown in PDF user manual only.

References

Burke, S. M. (2003). RTF Pocket Guide. " O'Reilly Media, Inc.".